+++ inc/Admin/advanced-settings.php
@@ -26,6 +26,8 @@ function adminAdvancedSettings()
             [
                 'crowdsec_stream_mode',
                 'crowdsec_stream_mode_refresh_frequency',
+                'crowdsec_usage_metrics_enabled',
+                'crowdsec_push_usage_metrics_frequency',
                 'crowdsec_redis_dsn',
                 'crowdsec_memcached_dsn',
                 'crowdsec_cache_system',
@@ -81,7 +83,7 @@ function adminAdvancedSettings()
      ** Section "Stream mode" **
      **************************/
     $streamMode = is_multisite() ? get_site_option('crowdsec_stream_mode') : get_option('crowdsec_stream_mode');
-    add_settings_section('crowdsec_admin_advanced_stream_mode', 'Communication mode to the API', function () {
+    add_settings_section('crowdsec_admin_advanced_stream_mode', 'Communication mode with Local API', function () {
     }, 'crowdsec_advanced_settings',['after_section' => '<hr>']);
 
     // Field "crowdsec_stream_mode"
@@ -144,6 +146,60 @@ function adminAdvancedSettings()
     '<a href="https://developer.wordpress.org/plugins/cron/hooking-wp-cron-into-the-system-task-scheduler/" target="_blank">'.
     'Here is explained how</a>.</p>', '...', 'width: 115px;', 'number');
 
+
+    /***************************
+     ** Section "Usage Metrics" **
+     **************************/
+    $isUsageMetricsEnabled = is_multisite() ? get_site_option('crowdsec_usage_metrics_enabled') : get_option('crowdsec_usage_metrics_enabled');
+    add_settings_section('crowdsec_admin_advanced_usage_metrics', 'Usage metrics', function () {
+    }, 'crowdsec_advanced_settings',['after_section' => '<hr>']);
+
+    // Field "crowdsec_usage_metrics_enabled"
+    addFieldCheckbox('crowdsec_usage_metrics_enabled', 'Enable Usage Metrics', 'crowdsec_plugin_advanced_settings', 'crowdsec_advanced_settings', 'crowdsec_admin_advanced_usage_metrics', function () {
+        // Usage metrics just activated.
+        scheduleUsageMetricsPush();
+    }, function () {
+        // Stream mode just deactivated.
+        unscheduleUsageMetricsPush();
+    }, '
+    <p>Enable usage metrics to gain visibility: monitor incoming traffic and blocked threats for better security insights.</p>
+    <p>For more information about usage metrics, please refer to the <a href="https://doc.crowdsec.net/docs/next/observability/usage_metrics/" target="_blank">documentation</a>.</p>
+    '.
+       ($isUsageMetricsEnabled ?
+           '<p><input id="crowdsec_push_usage_metrics" style="margin-right:10px" type="button" value="Push usage metrics now" class="button button-secondary button-small" onclick="document.getElementById(\'crowdsec_action_push_usage_metrics\').submit();"></p>' :
+           '<p><input id="crowdsec_push_usage_metrics" style="margin-right:10px" type="button" disabled="disabled" value="Push usage metrics now" class="button button-secondary button-small"></p>'));
+
+
+    // Field "crowdsec_push_usage_metrics_frequency"
+    addFieldString('crowdsec_push_usage_metrics_frequency', 'Push usage metrics each', 'crowdsec_plugin_advanced_settings', 'crowdsec_advanced_settings', 'crowdsec_admin_advanced_usage_metrics', function ($input) {
+        $input = (int) $input;
+        if ($input < 1) {
+            $input = 1;
+            $message = 'The "Push usage metrics each" value should be more than 1 sec (WP_CRON_LOCK_TIMEOUT). We just reset the frequency to 1 seconds.';
+            if(is_multisite()){
+                AdminNotice::displayError($message);
+            }else{
+                add_settings_error('Push usage metrics each', 'crowdsec_error', $message);
+            }
+
+            return $input;
+        }
+        $isUsageMetricsEnabled = is_multisite() ? get_site_option('crowdsec_usage_metrics_enabled') : get_option('crowdsec_usage_metrics_enabled');
+        // Update wp-cron schedule.
+        if ($isUsageMetricsEnabled) {
+            scheduleUsageMetricsPush();
+        }
+
+        return $input;
+    }, ' seconds. <p>Our advice is 60 seconds (as WordPress ignores durations under this value <a href="https://wordpress.stackexchange.com/questions/100104/better-handling-of-wp-cron-server-load-abuse" target="_blank">see WP_CRON_LOCK_TIMEOUT</a>).<br>'.
+       ' If you prefer a shorter delay between each push, you can <strong>go down to 1 sec</strong>.<br>'.
+       ' But as mentionned is the WordPress Developer Documentation, you should considere hooking WP-Cron Into the System Task Scheduler'.
+       ' by yourself and reduce the WP_CRON_LOCK_TIMEOUT value to the same value as you set here. '.
+       '<a href="https://developer.wordpress.org/plugins/cron/hooking-wp-cron-into-the-system-task-scheduler/" target="_blank">'.
+       'Here is explained how</a>.</p>', '...', 'width: 115px;', 'number');
+
+
+
     /*********************
      ** Section "Cache" **
      ********************/
@@ -205,20 +261,26 @@ function adminAdvancedSettings()
             $message = 'Technology: Incorrect cache technology selected.';
             if(is_multisite()){
                 AdminNotice::displayError($message);
-            }else{
+            } else{
                 add_settings_error('Technology', 'crowdsec_error', $message);
-
             }
         }
         $error = false;
 
         try {
             $configs = getDatabaseConfigs();
+            $isUsageMetricsEnabled = is_multisite() ? get_site_option('crowdsec_usage_metrics_enabled') : get_option('crowdsec_usage_metrics_enabled');
             $oldCacheSystem = $configs['crowdsec_cache_system'] ?? Constants::CACHE_SYSTEM_PHPFS;
             $bouncer = new Bouncer($configs);
+            if ($isUsageMetricsEnabled) {
+                $bouncer->pushUsageMetrics(Constants::BOUNCER_NAME, Constants::VERSION);
+            }
             $bouncer->clearCache();
             $message =
-                __('Cache system changed. Previous cache (' . $oldCacheSystem . ') data has been cleared. ');
+                __('Cache system changed.<br>Previous cache (' . $oldCacheSystem . ') data has been cleared. ');
+            if($isUsageMetricsEnabled){
+                $message .= '<br>As usage metrics push is enabled, metrics have been pushed before clearing cache.';
+            }
             AdminNotice::displaySuccess($message);
         } catch (Exception $e) {
             if (isset($configs['crowdsec_cache_system'])) {