+++ inc/templates/advanced-settings.php
@@ -30,6 +30,20 @@
         const $streamMode = jQuery('[name=crowdsec_stream_mode]');
         const $streamModeRefreshFrequency = jQuery('[name=crowdsec_stream_mode_refresh_frequency]');
         const $streamModeRefreshFrequencyTr = $streamModeRefreshFrequency.parent().parent();
+        // Usage Metrics
+        const $usageMetricsEnabled = jQuery('[name=crowdsec_usage_metrics_enabled]');
+        const $pushUsageMetricsFrequency = jQuery('[name=crowdsec_push_usage_metrics_frequency]');
+        const $pushUsageMetricsFrequencyTr = $pushUsageMetricsFrequency.parent().parent();
+
+        function updateUsageMetricsDisplay () {
+            if($usageMetricsEnabled.is(":checked")) {
+                $pushUsageMetricsFrequency.attr('required', 'required');
+                $pushUsageMetricsFrequencyTr.show("slow");
+            } else {
+                $pushUsageMetricsFrequency.removeAttr('required');
+                $pushUsageMetricsFrequencyTr.hide();
+            }
+        }
 
         function updateStreamModeDisplay () {
             if($streamMode.is(":checked")) {
@@ -98,10 +112,12 @@
 		updateDsnDisplay();
         updateAppSecDisplay();
         updateGeolocationDisplay();
+        updateUsageMetricsDisplay();
         $streamMode.change(updateStreamModeDisplay);
 		$cacheTechno.change(updateDsnDisplay);
         $useAppSec.change(updateAppSecDisplay);
         $geolocationEnabled.change(updateGeolocationDisplay);
+        $usageMetricsEnabled.change(updateUsageMetricsDisplay);
 	});
 	</script>
 	<?php settings_errors(); ?>
@@ -134,6 +150,10 @@
 			<input type="hidden" name="action" value="crowdsec_prune_cache">
 			<input type="hidden" name="nonce" value="<?php echo wp_create_nonce('crowdsec_prune_cache'); ?>">
 		</form>
+        <form action="<?php echo admin_url('admin-post.php'); ?>" method="post" id="crowdsec_action_push_usage_metrics">
+            <input type="hidden" name="action" value="crowdsec_push_usage_metrics">
+            <input type="hidden" name="nonce" value="<?php echo wp_create_nonce('crowdsec_push_usage_metrics'); ?>">
+        </form>
 		</div>
 		<p style="padding-top:15px">
 			Feel free to ask any questions about this plugin, make your suggestions or raise issues on the <a href="https://wordpress.org/support/plugin/crowdsec/">plugin support page</a> or directly on <a href="https://github.com/crowdsecurity/cs-wordpress-bouncer/issues/new">Github</a>.